+++ lib/yelo_icons.dart
@@ -14,10 +14,15 @@ class Yelo {
     Alignment alignment = Alignment.center, // New prop
     bool matchTextDirection = false, // New prop
     WidgetBuilder? placeholderBuilder, // New prop
+    VoidCallback? onTap, // New interaction prop
+    VoidCallback? onPress, // New interaction prop
+    VoidCallback? onLongPress, // New interaction prop
+    BorderRadius? borderRadius, // For ripple effect
+    bool enableFeedback = true, // Enable feedback sound/vibration
   }) {
     final String svgPath = '$_basePath/$iconName.svg';
 
-    return SvgPicture.asset(
+    final iconWidget = SvgPicture.asset(
       svgPath,
       width: size,
       height: size,
@@ -30,5 +35,18 @@ class Yelo {
               const Icon(Icons.error), // Using the new prop
       package: 'yelo_icons',
     );
+
+    if (onTap != null || onPress != null || onLongPress != null) {
+      return InkWell(
+        borderRadius: borderRadius ?? BorderRadius.circular(4), // Optional ripple effect
+        onTap: onTap ?? onPress, // Trigger onTap or onPress
+        onLongPress: onLongPress, // Trigger long press
+        enableFeedback: enableFeedback, // Enable feedback sound/vibration
+        child: iconWidget,
+      );
+    }
+
+    // Return the icon without gestures if no interaction is required
+    return iconWidget;
   }
 }